[UPDATE] refactored hooks, consistency improvement and included mail_branding to container_install manifest
