+++ mail_branding/__manifest__.py
@@ -16,5 +16,4 @@
     ],
     "installable": True,
     "uninstall_hook": "uninstall_hook",
-    "auto_install": True,
 }
