+++ mail_branding/models/res_users.py
@@ -5,7 +5,7 @@ class ResUsersPreferences(models.Model):
     _inherit = "res.users"
 
     notification_type = fields.Selection(
-        selection_add=[("inbox", "In CURQ")],
+        selection_add=[("inbox", "Handle in CURQ")],
         help="Policy on how to handle Chatter notifications:\n"
         "- By Emails: notifications are sent to your email address\n"
         "- In CURQ: notifications appear in your CURQ Inbox",
